+++ src/pages/customize/step1/index.tsx
@@ -6,8 +6,7 @@ import styled from 'styled-components'
 import { v4 as uuidv4 } from 'uuid'
 
 import { CancelCircle, DownArrow, HelpCircle, Plus, PlusBlack, LinkPrimary } from '@/assets/icons'
-import { useTracklistValidate } from '@/features/customize/model/useCustomize'
-import type { YoutubeVideoInfo } from '@/features/customize/types/customize'
+import { useTracklistValidate, type YoutubeVideoInfo } from '@/features/customize'
 import type { CustomizeStepProps } from '@/pages/customize'
 import { StepHeader, TrackItem, ToggleSwitch } from '@/pages/customize/ui'
 import { MUSIC_GENRES } from '@/shared/config/musicGenres'
